Added keys permission check for groups.all permissions, fix infinite loop in sync.
The infinite loop was caused by observing changes, which requested new syncs. In case of error of item update submission its `rawChangedFields` was set to some value >0 and the object was updated during sync process (some sync flag). This change would be observed by `RealmObjectChangeObserver` because there was an updated item with `rawChangedFields > 0`. So I added a new flag to all updatable items suggesting whether this change is made by user or sync process and then just observe the new flag.